+++ test/test.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"crypto/md5"
-	_ "errors"
 	"fmt"
 	"io"
 )
@@ -12,10 +11,10 @@ const (
 )
 
 func main() {
-	//c := [...]int{4, 5, 6}
-	//d := `hello
-	//world`
-	//fmt.Printf("%v,%v,%s",name,c,d)
+	c := [...]int{4, 5, 6}
+	d := `hello
+	world`
+	fmt.Printf("%v,%v,%s,%d",name,c,d,len(c))
 	//err := errors.New("emit macho dwarf: elf header corrupted")
 	//if err != nil  {
 	//	fmt.Println(err)
@@ -31,7 +30,7 @@ func main() {
 	var arr [10]int // 声明了一个int类型的数组
 	arr[0] = 42     // 数组下标是从0开始的
 	arr[1] = 13     // 赋值操作
-	c := [...]int{4, 5, 6}
+	//c := [...]int{4, 5, 6}
 	fmt.Printf("The first element is %d\n", arr[0]) // 获取数据，返回42
 	fmt.Printf("The last element is %d\n", arr[9])
 	fmt.Println(c)
